Fix BufferLength type to match original Thrift schema
Changed BufferLength from int? to sbyte? to match Int8Type from original
Thrift schema. The original code never synthesized BufferLength values -
it just passed through the null value from the Thrift server.

Changes:
- ColumnMetadataRecord.cs: Changed BufferLength from int? to sbyte?
- MetadataSchemaBuilder.cs: Changed AppendOrNull for Int8 to take sbyte?
  and removed unnecessary range validation (compiler enforces range)
- MetadataFieldPopulator.cs: Set BufferLength to null instead of calling
  GetBufferLength() (matches original behavior)

Verification:
- Test passes with all BUFFER_LENGTH values as null (matches baseline)
- Type is now byte-identical to original Thrift Int8Type
- Simpler code without runtime validation overhead

+++ csharp/src/Drivers/Apache/Hive2/Metadata/ColumnMetadataRecord.cs
@@ -71,8 +71,9 @@ namespace Apache.Arrow.Adbc.Drivers.Apache.Hive2.Metadata
         /// <summary>
         /// Transfer size of the data in bytes (BUFFER_LENGTH in statement-based metadata)
         /// Nullable for most types, only applicable for certain binary types
+        /// Uses sbyte to match Int8Type from original Thrift schema
         /// </summary>
-        public int? BufferLength { get; set; }
+        public sbyte? BufferLength { get; set; }
 
         /// <summary>
         /// Number of fractional digits (DECIMAL_DIGITS in statement-based metadata)
